INTLY-399 Open Che walkthrough resource in new tab in Walkthrough 2
Currently in Walkthrough 2 we have a walkthrough resource for the
Che console. When clicking on the link through the web app it opens
the link in the same tab that the user is currently in, meaning they
are taken away from the web app.

This change opens the Che console in a new tab so that the user
isn't completely taken away from their current context.