INTLY-1362 Add managedServices section to walkthrough.json
The web application currently provisions AMQ Online and Fuse Online
"slices" when the user firsts logs in. This is being changed to
instead provision those services as they are needed in walkthroughs.

Because of this change, we need a way to define the managed/shared
services from which "slices" are required. The web app will then
read these services and provision them in a way that is usable in
walkthroughs.

Only walkthrough 1A currently requires any "slices" of these
services, this updates 1A to include specifying "slices" of AMQ
Online and Fuse Online.